requirements.txt
<<<<<<< HEAD
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# bot\skill\caption\caption.py: 22
# bot\skill\caption\source\caption.py: 11
# bot\test.py: 12
Pillow == 6.1.0

# bot\skill\caption\source\train.py: 8
asposebarcode == 1.0.0

# bot\skill\caption\source\train.py: 8
asposestorage == 1.0.2

# bot\skill\caption\source\train.py: 8
automium-web == 0.1.1

# bot\bot.py: 21
command == 0.1.0

# bot\skill\caption\source\eval.py: 5
# bot\skill\caption\source\train.py: 9
datasets == 0.0.9

# bot\skill\caption\source\datasets.py: 3
# bot\skill\caption\source\utils.py: 3
h5py == 2.9.0

# bot\skill\caption\source\train.py: 8
hematopy == 0.0.1.dev6

# bot\skill\stylize\stylize.py: 13
# bot\stylize\neural_style_transfer.py: 9
imageio == 2.5.0

# bot\skill\stylize\stylize.py: 14
# bot\stylize\neural_style_transfer.py: 10
# bot\stylize\neural_style_transfer_examine.py: 7
# bot\stylize\neural_style_transfer_video.py: 5,6,9
imutils == 0.5.3

# bot\skill\caption\source\train.py: 8
indras-net == 2.0.6

# bot\joke\deepjoke.py: 12
keras == 2.2.5

# bot\skill\caption\caption.py: 18,19
# bot\skill\caption\source\caption.py: 6,7
matplotlib == 3.1.1

# bot\skill\caption\source\eval.py: 7
# bot\skill\caption\source\train.py: 11
nltk == 3.4.5

# bot\joke\deepjoke.py: 14
# bot\joke\joke.py: 8
# bot\kmeans\kmeans.py: 2
# bot\mnist\load_data.py: 8
# bot\mnist\mnist.py: 11
# bot\skill\caption\caption.py: 16
# bot\skill\caption\source\caption.py: 3
# bot\skill\caption\source\utils.py: 2
# bot\skill\kmeans\kmeans.py: 10
numpy == 1.16.4

# bot\skill\skill.py: 11
# bot\skill\stylize\stylize.py: 15
# bot\stylize\neural_style_transfer.py: 11
# bot\stylize\neural_style_transfer_examine.py: 8
# bot\stylize\neural_style_transfer_video.py: 11
opencv-contrib-python == 4.1.0.25

# bot\skill\skill.py: 11
# bot\skill\stylize\stylize.py: 15
# bot\stylize\neural_style_transfer.py: 11
# bot\stylize\neural_style_transfer_examine.py: 8
# bot\stylize\neural_style_transfer_video.py: 11
opencv-contrib-python-headless == 4.1.0.25

# bot\skill\skill.py: 11
# bot\skill\stylize\stylize.py: 15
# bot\stylize\neural_style_transfer.py: 11
# bot\stylize\neural_style_transfer_examine.py: 8
# bot\stylize\neural_style_transfer_video.py: 11
opencv-python == 4.1.0.25

# bot\skill\skill.py: 11
# bot\skill\stylize\stylize.py: 15
# bot\stylize\neural_style_transfer.py: 11
# bot\stylize\neural_style_transfer_examine.py: 8
# bot\stylize\neural_style_transfer_video.py: 11
opencv-python-aarch64 == 3.3.0.1

# bot\skill\skill.py: 11
# bot\skill\stylize\stylize.py: 15
# bot\stylize\neural_style_transfer.py: 11
# bot\stylize\neural_style_transfer_examine.py: 8
# bot\stylize\neural_style_transfer_video.py: 11
opencv-python-headless == 4.1.0.25

# bot\bot.py: 20
pygsl == 2.3.0

# bot\skill\caption\source\train.py: 8
pynotion == 0.0.1

# bot\skill\caption\source\train.py: 8
pyskyscanner == 1.0.1

# bot\bot.py: 15
# bot\skill\caption\caption.py: 15
requests == 2.22.0

# bot\skill\caption\source\train.py: 8
roomai == 0.1.16

# bot\skill\caption\caption.py: 17
# bot\skill\caption\source\caption.py: 8
scikit_image == 0.15.0

# bot\kmeans\kmeans.py: 3
# bot\mnist\mnist.py: 12
# bot\skill\caption\source\caption.py: 10
# bot\skill\caption\source\utils.py: 6
# bot\skill\kmeans\kmeans.py: 11
scipy == 1.3.1

# bot\bot.py: 17
slackclient == 2.1.0

# bot\skill\caption\source\train.py: 8
solar-radiation-model == 0.0.20

# bot\skill\caption\source\train.py: 8
swiftcodegen == 0.3

# bot\skill\caption\caption.py: 12,20
# bot\skill\caption\models.py: 3,4
# bot\skill\caption\source\caption.py: 1,2
# bot\skill\caption\source\datasets.py: 1,2
# bot\skill\caption\source\eval.py: 1,2,3,8
# bot\skill\caption\source\models.py: 3,4
# bot\skill\caption\source\train.py: 2,3,4,6,7
# bot\skill\caption\source\utils.py: 5
torch == 1.2.0

# bot\skill\caption\caption.py: 21
# bot\skill\caption\models.py: 5
# bot\skill\caption\source\caption.py: 5
# bot\skill\caption\source\eval.py: 4
# bot\skill\caption\source\models.py: 5
# bot\skill\caption\source\train.py: 5
torchvision == 0.4.0

# bot\skill\caption\source\eval.py: 9
# bot\skill\caption\source\utils.py: 7
tqdm == 4.35.0

# bot\skill\caption\source\train.py: 8
transferflow == 0.1.8

# bot\skill\caption\source\train.py: 8
visvmtagger == 1.0a0
=======
scipy==1.1.0
PyMarkovChain==1.8
imageio==2.4.1
imutils==0.5.2
slackclient==1.3.0
Keras==2.2.4
numpy==1.15.4
requests==2.20.0
opencv_contrib_python==4.0.0.21
>>>>>>> ee4c0adbd64cd429d5ef34d403aa4480ae115c97
